requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
python-jose[cryptography]
argon2-cffi
python-multipart
python-dotenv
pydantic  
pathlib
pydantic[email]
cloudinary